# docs/openapi.json
{
  "components": {
    "schemas": {
      "CreateWalletRequest": {
        "additionalProperties": false,
        "properties": {
          "initialBalance": {
            "additionalProperties": false,
            "properties": {
              "amount": {
                "pattern": "^\\d+\\.\\d{2}$",
                "type": "string"
              },
              "currency": {
                "pattern": "^[A-Z]{3}$",
                "type": "string"
              }
            },
            "required": [
              "amount",
              "currency"
            ],
            "type": "object"
          },
          "playerId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          }
        },
        "required": [
          "playerId",
          "initialBalance"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "failureCode": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "failureCode",
          "message"
        ],
        "type": "object"
      },
      "LedgerPageResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "additionalProperties": false,
              "properties": {
                "amount": {
                  "additionalProperties": false,
                  "properties": {
                    "amount": {
                      "pattern": "^\\d+\\.\\d{2}$",
                      "type": "string"
                    },
                    "currency": {
                      "pattern": "^[A-Z]{3}$",
                      "type": "string"
                    }
                  },
                  "required": [
                    "amount",
                    "currency"
                  ],
                  "type": "object"
                },
                "balanceAfter": {
                  "additionalProperties": false,
                  "properties": {
                    "amount": {
                      "pattern": "^\\d+\\.\\d{2}$",
                      "type": "string"
                    },
                    "currency": {
                      "pattern": "^[A-Z]{3}$",
                      "type": "string"
                    }
                  },
                  "required": [
                    "amount",
                    "currency"
                  ],
                  "type": "object"
                },
                "balanceBefore": {
                  "additionalProperties": false,
                  "properties": {
                    "amount": {
                      "pattern": "^\\d+\\.\\d{2}$",
                      "type": "string"
                    },
                    "currency": {
                      "pattern": "^[A-Z]{3}$",
                      "type": "string"
                    }
                  },
                  "required": [
                    "amount",
                    "currency"
                  ],
                  "type": "object"
                },
                "createdAt": {
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
                  "type": "string"
                },
                "direction": {
                  "enum": [
                    "CREDIT",
                    "DEBIT"
                  ],
                  "type": "string"
                },
                "entryHash": {
                  "pattern": "^[a-f0-9]{64}$",
                  "type": "string"
                },
                "entrySequence": {
                  "exclusiveMinimum": 0,
                  "maximum": 9007199254740991,
                  "type": "integer"
                },
                "id": {
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "type": "string"
                },
                "previousEntryHash": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "transactionId": {
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "type": "string"
                },
                "walletId": {
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "type": "string"
                }
              },
              "required": [
                "id",
                "walletId",
                "transactionId",
                "entrySequence",
                "direction",
                "amount",
                "balanceBefore",
                "balanceAfter",
                "previousEntryHash",
                "entryHash",
                "createdAt"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object"
      },
...
      "ReconciliationResponse": {
        "additionalProperties": false,
        "properties": {
          "accountingBalance": {
            "additionalProperties": false,
            "properties": {
              "amount": {
                "pattern": "^\\d+\\.\\d{2}$",
                "type": "string"
              },
              "currency": {
                "pattern": "^[A-Z]{3}$",
                "type": "string"
              }
            },
            "required": [
              "amount",
              "currency"
            ],
            "type": "object"
          },
          "accountingBalanced": {
            "type": "boolean"
          },
          "auditChainValid": {
            "type": "boolean"
          },
          "calculatedBalance": {
            "additionalProperties": false,
            "properties": {
              "amount": {
                "pattern": "^\\d+\\.\\d{2}$",
                "type": "string"
              },
              "currency": {
                "pattern": "^[A-Z]{3}$",
                "type": "string"
              }
            },
            "required": [
              "amount",
              "currency"
            ],
            "type": "object"
          },
          "checkedEntries": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          },
          "consistent": {
            "type": "boolean"
          },
          "difference": {
            "additionalProperties": false,
            "properties": {
              "amount": {
                "pattern": "^-?\\d+\\.\\d{2}$",
                "type": "string"
              },
              "currency": {
                "pattern": "^[A-Z]{3}$",
                "type": "string"
              }
            },
            "required": [
              "amount",
              "currency"
            ],
            "type": "object"
          },
          "storedBalance": {
            "additionalProperties": false,
            "properties": {
              "amount": {
                "pattern": "^\\d+\\.\\d{2}$",
                "type": "string"
              },
              "currency": {
                "pattern": "^[A-Z]{3}$",
                "type": "string"
              }
            },
            "required": [
              "amount",
              "currency"
            ],
            "type": "object"
          },
          "walletId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          }
        },
        "required": [
          "walletId",
          "storedBalance",
          "calculatedBalance",
          "accountingBalance",
          "difference",
          "consistent",
          "checkedEntries",
          "accountingBalanced",
          "auditChainValid"
        ],
        "type": "object"
      },
...
      "WalletResponse": {
        "additionalProperties": false,
        "properties": {
          "balance": {
            "additionalProperties": false,
            "properties": {
              "amount": {
                "pattern": "^\\d+\\.\\d{2}$",
                "type": "string"
              },
              "currency": {
                "pattern": "^[A-Z]{3}$",
                "type": "string"
              }
            },
            "required": [
              "amount",
              "currency"
            ],
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "playerId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "version": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "playerId",
          "balance",
          "version"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "HTTP contract for wallet, wagering, reconciliation, and operational endpoints.",
    "title": "Distributed Wagering Processor API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/wallets": {
      "post": {
        "operationId": "createWallet",
        "parameters": [
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "initialBalance": {
                  "amount": "100.00",
                  "currency": "BRL"
                },
                "playerId": "22222222-2222-4222-8222-222222222222"
              },
              "schema": {
                "$ref": "#/components/schemas/CreateWalletRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "balance": {
                    "amount": "100.00",
                    "currency": "BRL"
                  },
                  "id": "11111111-1111-4111-8111-111111111111",
                  "playerId": "22222222-2222-4222-8222-222222222222",
                  "version": 1
                },
                "schema": {
                  "$ref": "#/components/schemas/WalletResponse"
                }
              }
            },
            "description": "Wallet created"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "failureCode": "INVALID_PAYLOAD",
                  "message": "Request payload is invalid"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Wallet already exists"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Dependency temporarily unavailable"
          }
        },
        "summary": "Create a wallet and its opening financial records",
        "tags": [
          "Wallets"
        ]
      }
    },
    "/wallets/{walletId}": {
      "get": {
        "operationId": "getWallet",
        "parameters": [
          {
            "description": "Wallet UUID",
            "in": "path",
            "name": "walletId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "balance": {
                    "amount": "100.00",
                    "currency": "BRL"
                  },
                  "id": "11111111-1111-4111-8111-111111111111",
                  "playerId": "22222222-2222-4222-8222-222222222222",
                  "version": 1
                },
                "schema": {
                  "$ref": "#/components/schemas/WalletResponse"
                }
              }
            },
            "description": "Wallet found"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "failureCode": "INVALID_PAYLOAD",
                  "message": "Request payload is invalid"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Wallet not found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Dependency temporarily unavailable"
          }
        },
        "summary": "Get a wallet",
        "tags": [
          "Wallets"
        ]
      }
    },
    "/wallets/{walletId}/ledger": {
      "get": {
        "operationId": "listWalletLedger",
        "parameters": [
          {
            "description": "Wallet UUID",
            "in": "path",
            "name": "walletId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "items": [
                    {
                      "amount": {
                        "amount": "100.00",
                        "currency": "BRL"
                      },
                      "balanceAfter": {
                        "amount": "100.00",
                        "currency": "BRL"
                      },
                      "balanceBefore": {
                        "amount": "0.00",
                        "currency": "BRL"
                      },
                      "createdAt": "2026-01-01T00:00:00.000Z",
                      "direction": "CREDIT",
                      "entryHash": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                      "entrySequence": 1,
                      "id": "44444444-4444-4444-8444-444444444444",
                      "previousEntryHash": null,
                      "transactionId": "33333333-3333-4333-8333-333333333333",
                      "walletId": "11111111-1111-4111-8111-111111111111"
                    }
                  ],
                  "nextCursor": null
                },
                "schema": {
                  "$ref": "#/components/schemas/LedgerPageResponse"
                }
              }
            },
            "description": "Ledger page"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "failureCode": "INVALID_PAYLOAD",
                  "message": "Request payload is invalid"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Wallet not found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Dependency temporarily unavailable"
          }
        },
        "summary": "List immutable ledger entries using a stable opaque cursor",
        "tags": [
          "Wallets"
        ]
      }
    },
    "/wallets/{walletId}/reconciliation": {
      "post": {
        "operationId": "reconcileWallet",
        "parameters": [
          {
            "description": "Wallet UUID",
            "in": "path",
            "name": "walletId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "accountingBalance": {
                    "amount": "100.00",
                    "currency": "BRL"
                  },
                  "accountingBalanced": true,
                  "auditChainValid": true,
                  "calculatedBalance": {
                    "amount": "100.00",
                    "currency": "BRL"
                  },
                  "checkedEntries": 1,
                  "consistent": true,
                  "difference": {
                    "amount": "0.00",
                    "currency": "BRL"
                  },
                  "storedBalance": {
                    "amount": "100.00",
                    "currency": "BRL"
                  },
                  "walletId": "11111111-1111-4111-8111-111111111111"
                },
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationResponse"
                }
              }
            },
            "description": "Reconciliation report"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "failureCode": "INVALID_PAYLOAD",
                  "message": "Request payload is invalid"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Wallet not found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Dependency temporarily unavailable"
          }
        },
        "summary": "Run a read-only full financial reconciliation",
        "tags": [
          "Wallets"
        ]
      }
    }
  },
...
}